fix segment pointer truncation (#15)

+++ mm/src/code/z_actor.c
@@ -3385,7 +3385,7 @@ Actor* Actor_SpawnAsChildAndCutscene(ActorContext* actorCtx, PlayState* play, s1
     Actor_AddToCategory(actorCtx, actor, actorInit->type);
 
     {
-        u32 sp20 = gSegments[6];
+        uintptr_t sp20 = gSegments[6];
 
         Actor_Init(actor, play);
         gSegments[6] = sp20;
